% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab2group2function.R
\name{load_DRG_data}
\alias{load_DRG_data}
\title{Load DRG Data from CSV}
\usage{
load_DRG_data()
}
\value{
A data frame containing the DRG data.
}
\description{
This function loads the DRG_data.csv file included in the package.
}
